+++ lib/uniword.rb
@@ -134,6 +134,7 @@ module Uniword
   autoload :CLI, "uniword/cli/main"
   autoload :GenerateCLI, "uniword/cli/generate_cli"
   autoload :FontsCLI, "uniword/cli/fonts_cli"
+  autoload :PageCLI, "uniword/cli/page_cli"
   autoload :DiffCLI, "uniword/cli/diff_cli"
   autoload :TemplateCLI, "uniword/cli/template_cli"
   autoload :CLIHelpers, "uniword/cli/helpers"